Refactored the input system completely. And inserted standard config handler
This mostly includes wrapper functions for the original ui functions
and a more dynamic (and perhaps also incomprhensible) method resolving
system. This was thought as being easily expandable and generic, but
only time will tell if this will work.

Also: as my mind does not have enough bits for processing the
complexity of the program as it runs, I have added a vast number of
debug statements.

A small setup routine was added, that tries to construct a greenfield
working environment.